(#311) - Example for the REDIS_URL for the redisdown
This is a quick and dirty example of where I put the redisdown REDIS_URL in so that one could use redis on
heroku or if your redis server is not on localhost.  This example is based on a 2016 hack that I did, and I
haven't proved it out on the most recent plugin based code.

Update index.js

Not all level backend databases are redis.